build issues for UT
